Cyclops Syncing Review Changes
- Removed server time syncing. In-game time syncing still working.
- Moved damage and fire logic that was only used in CyclopsDamageProcessor out of Cyclops.
- Removed header file from an aborted rebase because I have no idea what I'm doing.
- Added buffering to CyclopsFireHealthChanged packet.
